Create Computer_vision.py
Computer vision is a field of computer science that enables a device to efficiently and accurately “see” and “understand” the content present within a digital image, such as videos and photographs, as is done by humans. Artificial Intelligence (AI) is using computer vision mainly for image processing. In this lab, you will learn how to:
•	Install computer vision libraries
•	Read, write and display the input image
•	Perform color conversion and edge detection on the input image
•	Play an input video